inventory_management/requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==24.2.0
certifi==2024.7.4
cffi==1.17.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Django==5.0.6
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
fastapi==0.112.2
fonttools==4.53.1
h11==0.14.0
idna==3.7
imageio==2.34.2
kiwisolver==1.4.5
matplotlib==3.9.1
mysqlclient==2.2.4
numpy==2.0.1
opencv-python==4.10.0.84
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pefile==2023.2.7
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyinstaller==6.8.0
pyinstaller-hooks-contrib==2024.7
PyJWT==1.7.1
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32-ctypes==0.2.2
redis==5.1.0
selenium==4.23.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.0
starlette==0.38.4
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
websocket-client==1.8.0
wsproto==1.2.0
xlrd==2.0.1
